Extract detector models on a module
